Eliminate /health test code duplication (#1443)
